all features: generate rows in a loader
