+++ website/routes/AllFeatures.tsx
@@ -1,0 +1,38 @@
+import { faker } from '@faker-js/faker';
+import { createFileRoute } from '@tanstack/react-router';
+
+import type { Row } from './AllFeatures.lazy';
+
+let rows: readonly Row[] | undefined;
+
+function createRows(): Row[] {
+  const rows: Row[] = [];
+
+  for (let i = 0; i < 2000; i++) {
+    rows.push({
+      id: `id_${i}`,
+      avatar: faker.image.avatar(),
+      email: faker.internet.email(),
+      title: faker.person.prefix(),
+      firstName: faker.person.firstName(),
+      lastName: faker.person.lastName(),
+      street: faker.location.street(),
+      zipCode: faker.location.zipCode(),
+      date: faker.date.past().toLocaleDateString(),
+      bs: faker.company.buzzPhrase(),
+      catchPhrase: faker.company.catchPhrase(),
+      companyName: faker.company.name(),
+      words: faker.lorem.words(),
+      sentence: faker.lorem.sentence()
+    });
+  }
+
+  return rows;
+}
+
+export const Route = createFileRoute('/AllFeatures')({
+  loader() {
+    rows ??= createRows();
+    return rows;
+  }
+});
